SimpleClansEconomyTransactionEvent has been renamed to EconomyTransactionEvent chore!: Old methods PermissionsManager#playerGrantMoney(OfflinePlayer, double) & PermissionsManager#playerChargeMoney(OfflinePlayer, double) has been marked as deprecated and for removal. BREAKING CHANGE: Deprecated methods Clan#deposit & Clan#withdraw have been removed
