+++ src/main/java/net/sacredlabyrinth/phaed/simpleclans/commands/clan/BankCommand.java
@@ -16,7 +16,7 @@ import org.bukkit.entity.Player;
 import static net.sacredlabyrinth.phaed.simpleclans.SimpleClans.lang;
 import static net.sacredlabyrinth.phaed.simpleclans.events.ClanBalanceUpdateEvent.Cause.COMMAND;
 import static net.sacredlabyrinth.phaed.simpleclans.events.ClanBalanceUpdateEvent.Cause.REVERT;
-import static net.sacredlabyrinth.phaed.simpleclans.events.SimpleClansEconomyTransactionEvent.Cause.CLAN_BANK;
+import static net.sacredlabyrinth.phaed.simpleclans.events.EconomyTransactionEvent.Cause.CLAN_BANK;
 import static org.bukkit.ChatColor.AQUA;
 import static org.bukkit.ChatColor.RED;
